requirements.txt
#dependecies for project

yfinance>=0.2.32
newsapi-python>=0.2.7
pandas>=2.1.0
numpy>=1.24.0
pandas-ta>=0.3.14b
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
